Folder structure update: reorganize components to features folder
