+++ src/features/community/events/CarouselForThumbnail.tsx
