+++ src/features/community/main/CarouselWithDots.tsx
