+++ src/features/community/main/EventBanner.tsx
