+++ src/features/community/main/EventBannerSlider.tsx
