+++ src/features/community/main/EventPostSlider.tsx
