+++ src/features/community/main/PloggingPostSlider.tsx
@@ -4,7 +4,7 @@ import React, { useCallback } from 'react';
 import Image from 'next/image';
 import { PloggingPost } from '@/services/ploggingPosts';
 import Avatar from '../../public/images/avatar.png';
-import LottieHeart from './icon/LottieHeart';
+import LottieHeart from '../../../components/icon/LottieHeart';
 import useIntersectionObserver from '@/hooks/useIntersectionObserver';
 
 type Props = { posts: PloggingPost[] };
