+++ src/features/community/main/RecruitPostSlider.tsx
